+++ modules/portable/SaveToFile/list_pwd.h
@@ -15,5 +15,5 @@ class ListPwd
 		static void appendEntries(const char* dir, unsigned int entry_types, unsigned int flags, std::vector<ppk_entry*>& list);
 	public:
 		static unsigned int getEntryListCount(const char* dir, unsigned int entry_types, unsigned int flags);
-		static ppk_entry** getEntryList(const char* dir, unsigned int entry_types, unsigned int* nbEntries, unsigned int flags);
+		static ppk_entry** getEntryList(const char* dir, unsigned int entry_types, size_t* nbEntries, unsigned int flags);
 };
